Separate tests into two files
